remove debug logger
